Abstracted Image Memory management

+++ src/filters.h
@@ -4,4 +4,4 @@
 // Functions to manipulate images
 NormImage* NormalizeRawImage(RawImage* inputImage);
 NormImage* RgbToPrintable(RawImage* inputImage);
-NormImage* RadonTransform(NormImage* inputimage, float angles[], int nangles, int nbins, int color, int resolution);
+NormImage* RadonTransform(NormImage* inputimage, double angles[], int nangles, int nbins, int color, int resolution);
